+++ VORP-Inventory[Client-Server]/vorpinventory_sv/ItemDatabase.cs
@@ -3,6 +3,7 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
+using vorpinventory_sv.Models;
 
 namespace vorpinventory_sv
 {
